hop-node: Vault: Add interface
